Initial Release
Added code to check if the user making the request can access the game.

+++ TbspRpgApi/Controllers/GamesController.cs
@@ -37,6 +37,9 @@ namespace TbspRpgApi.Controllers
         public async Task<IActionResult> DeleteGame(Guid gameId)
         {
             // make sure the user either is the game's owner or the user is an admin
+            if(!CanAccessGame(gameId))
+                return BadRequest(new { message = NotYourGameErrorMessage });
+            
             try
             {
                 await _gamesService.DeleteGame(gameId);
